+++ src/lotus_json/mod.rs
@@ -213,6 +213,7 @@ decl_and_test!(
     vrf_proof for crate::blocks::VRFProof,
 );
 
+mod beneficiary_term; // fil_actor_miner_state::v12::BeneficiaryTerm: !quickcheck::Arbitrary
 mod bit_field; //  fil_actors_shared::fvm_ipld_bitfield::BitField: !quickcheck::Arbitrary
 mod cid; // can't make snapshots of generic type
 mod claim; // fil_actor_interface::power::Claim: !quickcheck::Arbitrary
@@ -220,6 +221,7 @@ mod hash_map;
 mod ipld; // NaN != NaN
 mod nonempty;
 mod opt; // can't make snapshots of generic type
+mod pending_beneficiary_change; // fil_actor_miner_state::v12::PendingBeneficiaryChange: !quickcheck::Arbitrary
 mod raw_bytes; // fvm_ipld_encoding::RawBytes: !quickcheck::Arbitrary
 mod receipt; // shim type roundtrip is wrong - see module
 mod vec; // can't make snapshots of generic type
